weather_update/pyproject.toml
[project]
name = "weather-update"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "aleema", email = "aleemasaleem416324@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "openai>=0.27.10",
    "openai-agents>=0.2.4",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.10",
]

[project.scripts]
weather-update = "weather_update:main"

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"
